Collection of improvements and version 0.3 release
No new features, but a collection of improvements to improve stability.
There are enough changes to warrant a minor version bump. Changes
include:

- Don't subscribe to windowMoved window filter event. Instead create an
UI element watcher for each window. Store the watchers in a table.
Create a moveWindow() function that will pause the UI element watcher
for just that window when setting a new frame. This lets us avoid move
events from the window we know we're moving without throwing way move
events from all the other windows in the window filter.

- The windowAllowed in filter event is too broad. Switch to
windowVisiblt, windowNotVisible, windowFullscreened, and
windowUnfullscreened events. This is enough to cover windows created,
destroyed, hidden, minimized, and fullscreened. Don't try to detect when
a window moves to a new space with windowNotInCurrentSpace. This doesn't
work well.

- Use the windowFocused event to store the previously focused window.
When a new window is created/visible the order of events are 1)
new windowVisible then 2) new windowFocused. We can use this order to
query the saved previously focused window and know where to insert the
new window. Store the actual hs.window userdata instead of the window ID
because not all windows (like Google Chrome progressive web apps) can be
looked up from their ID.

- Tile windows by space. For a given space, see if the currently focused
window is in the space. If not, find the first visible on-screen window.
Use either of these as the anchor window to adjust the other windows in
relation to.

- Use Lua 5.4 const variables.

- Add local hs imports to avoid global lookups at runtime.