Store a pending window until it has a Space and can be added
When a new Window is created, we first get a windowVisible event but
without a Space. Next we receive a windowFocused event for the Window,
but this also sometimes lacks a Space. Store the Window pending a Space
in the pending_window variable and set a timer to try to add the Window
again later. Also schedule the windowFocused handler to run later after
the window was added.

Merge the separate window event handlers into a unified handler.

Additionally, when a Window is destroyed, always try to focus on a
nearby Window. MacOS will jump focus to another Window in the same App
BEFORE issuing the windowNotVisible event. This other same App Window
may be on another screen, or far away, and is usually not what a user
would want.

This seems like a significant enough change to bump the version number.