Raise app before focusing window on new Space
Following a tip in this Hammerspoon issue thread:
https://github.com/Hammerspoon/hammerspoon/issues/370
Activate the app a window belongs to, wait a small amount, then try to
focus the window on a new Space. Do this before switching to the new
Space so a window is focused and PaperWM.spoon is ready to navigate via
keyboard shortcuts after switching.