Add a refresh_windows command
In case window tiling gets into a bad state (happens when attaching
and detaching an external monitor) provide a command to refresh the
current windows and retile spaces as required.

Collect all windows from the current window filter and check if they
exist in the window list and if there space has changed.